# api/openapi.yaml
openapi: 3.1.0
info:
  title: Domain Monitoring & Asset Intelligence API
  version: 0.6.0-phase6
  description: Bilingual Thai/English domain monitoring, Drive/Excel imports, rule-based recommendations and reporting API.
servers:
  - url: /api/v1
paths:
...
  /google-sheets/excel/previews:
    post:
      summary: Upload a bounded .xlsx workbook and persist an import preview
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: "#/components/parameters/CSRFToken"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
                source_name: {type: string}
                sheet_name: {type: string}
                column_mapping: {type: string, description: JSON object mapping semantic names to workbook headers}
      responses:
        "201": {description: Preview with source_kind excel}
        "413": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: domainintel_session
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 200}
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message, messages, locale, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            messages:
              type: object
              required: [th, en]
              properties:
                th: {type: string}
                en: {type: string}
            locale: {type: string, enum: [th, en]}
            request_id: {type: string, format: uuid}
            details: {type: object, additionalProperties: true}
  responses:
    Error:
      description: Bilingual error response
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}